add -all components use context data + Login form has 'logging in' indicator while loading

+++ src/services/deleteService.tsx
@@ -1,41 +1,43 @@
-const fetchDelete = (endpoint: string, itemId: string) => {
-  fetch(`http://localhost:4000/api/content/${endpoint}`, {
-    method: 'DELETE', // or 'PUT'
-    mode: 'cors',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({ "_id": itemId }),
-  })
-    .then(res => res.json())
-    .then(data => {
-      console.log('Success:', data);
+const fetchDelete = (endpoint: string, itemId: string, authToken: string) => {
+  window.confirm('Sure you want to delete this?') &&
+    fetch(`http://localhost:4000/api/content/${endpoint}`, {
+      method: 'DELETE', // or 'PUT'
+      mode: 'cors',
+      headers: {
+        'Content-Type': 'application/json',
+        'jwt': authToken
+      },
+      body: JSON.stringify({ "_id": itemId }),
     })
-    .catch((error) => {
-      console.error('Error:', error);
-    });
+      .then(res => res.json())
+      .then(data => {
+        console.log('Success:', data);
+      })
+      .catch((error) => {
+        console.error('Error:', error);
+      });
 }
 
-export const deleteItem = (itemId: string, itemType: ItemType) => {
+export const deleteItem = (itemId: string, itemType: ItemType, authToken: string) => {
   switch (itemType) {
     case 'about':
-      fetchDelete('about', itemId);
+      fetchDelete('about', itemId, authToken);
       break;
 
     case 'skill':
-      fetchDelete('skills', itemId);
+      fetchDelete('skills', itemId, authToken);
       break;
 
     case 'work':
-      fetchDelete('work', itemId);
+      fetchDelete('work', itemId, authToken);
       break;
 
     case 'studies':
-      fetchDelete('studies', itemId);
+      fetchDelete('studies', itemId, authToken);
       break;
 
     case 'project':
-      fetchDelete('projects', itemId);
+      fetchDelete('projects', itemId, authToken);
       break;
 
     default:
